changed files into packages, took off log out feature
