+++ opendb/opendb.go
@@ -1,4 +1,4 @@
-package main
+package opendb
 
 import (
 	"database/sql"
